move activation & deactivation classes
