+++ inc/Engine/Deactivation/Deactivation.php
@@ -1,12 +1,21 @@
 <?php
 
-namespace WP_Rocket;
+namespace WP_Rocket\Engine\Deactivation;
 
+use League\Container\Container;
 use WP_Rocket\Engine\Cache\AdvancedCache;
-use WP_Rocket\Engine\Cache\WPCache;
-use WP_Rocket\Engine\Capabilities\Manager;
 
 class Deactivation {
+    /**
+	 * Aliases in the container for each class that needs to call its deactivate method
+	 *
+	 * @var array
+	 */
+	private static $deactivators = [
+		'capabilities_manager',
+		'wp_cache',
+    ];
+
 	/**
 	 * Performs these actions during the plugin deactivation
 	 *
@@ -15,8 +24,14 @@ class Deactivation {
 	public static function deactivate() {
 		global $is_apache;
 
-        $filesystem = rocket_direct_filesystem();
-        $wp_cache   = new WPCache( $filesystem );
+        $container  = new Container();
+		$filesystem = rocket_direct_filesystem();
+
+		$container->addServiceProvider( 'WP_Rocket\Engine\Deactivation\ServiceProvider' );
+	
+		foreach ( self::$deactivators as $deactivator ) {
+			$container->get( $deactivator );
+		}
 
         if ( ! isset( $_GET['rocket_nonce'] ) || ! wp_verify_nonce( sanitize_key( $_GET['rocket_nonce'] ), 'force_deactivation' ) ) {
             $causes = [];
@@ -28,7 +43,7 @@ class Deactivation {
 
             // wp-config problem.
             if (
-                ! $wp_cache->find_wpconfig_path()
+                ! $container->get( 'wp_cache' )->find_wpconfig_path()
                 &&
                 // This filter is documented in inc/Engine/Cache/WPCache.php.
                 (bool) apply_filters( 'rocket_set_wp_cache_constant', true )
@@ -49,8 +64,6 @@ class Deactivation {
         if ( ! count( glob( WP_ROCKET_CONFIG_PATH . '*.php' ) ) ) {
             // Delete All WP Rocket rules of the .htaccess file.
             flush_rocket_htaccess( true );
-
-            $wp_cache->deactivate();
 
             // Delete content of advanced-cache.php.
             rocket_put_content( WP_CONTENT_DIR . '/advanced-cache.php', '' );
@@ -73,8 +86,6 @@ class Deactivation {
         wp_clear_scheduled_hook( 'rocket_facebook_tracking_cache_update' );
         wp_clear_scheduled_hook( 'rocket_google_tracking_cache_update' );
         wp_clear_scheduled_hook( 'rocket_cache_dir_size_check' );
-
-        ( new Manager() )->deactivate();
 
         /**
          * WP Rocket deactivation.
